src/requirements-alternative.txt
# Alternative requirements.txt using pypdf instead of PyMuPDF
# Use this if PyMuPDF installation fails

# Core framework
streamlit>=1.31.0

# Data models
pydantic>=2.6.0

# Document parsing - Alternative approach
pypdf>=4.0.0  # Simpler PDF library with pure Python implementation
python-docx>=1.1.0
pdfplumber>=0.10.0

# ML & Embeddings
sentence-transformers>=2.3.0
torch>=2.2.0

# Vector storage
faiss-cpu>=1.7.0

# Data processing
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.2.0

# Utilities
python-dotenv>=1.0.0

# Visualization
plotly>=5.18.0

# Async
aiofiles>=23.2.0

# Similarity metrics
scikit-learn>=1.3.0
